requirements.txt
langchain==0.1.0
chromadb==0.4.18
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
python-multipart==0.0.6
tiktoken==0.5.1
numpy==1.26.2
pandas==2.1.3
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0 